figured out duplicate provider issue
solution => remove callback onOpen,onDismiss, onComplete replace with event emitter that can be listened to
that way the state controlling widget can be below the config provider in the tree.
since we need to have callbacks in the config provider a state controlling widget would need to sit on top of the provider.
since the provider should be very close to the root the state controller would be forced to live there as well, and also provider the config provider instanc.
hence potentially introducing a second provider config implicitly ...